+++ src/main/java/com/overhere/backend/config/WebClientConfig.java
@@ -1,6 +1,5 @@
-package com.overhere.backend.configuration;
+package com.overhere.backend.config;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.reactive.ReactorClientHttpConnector;
