+++ src/main/java/com/overhere/backend/dao/NonObstacleInfoRepository.java
@@ -1,6 +1,6 @@
-package com.overhere.backend.repository;
+package com.overhere.backend.dao;
 
-import com.overhere.backend.entity.NonObstacleInfo;
+import com.overhere.backend.domain.NonObstacleInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
