+++ src/main/java/com/overhere/backend/domain/enums/AreaCode.java
@@ -1,0 +1,5 @@
+package com.overhere.backend.domain.enums;
+
+public enum AreaCode {
+    TEMP;
+}
