+++ src/main/java/com/overhere/backend/domain/enums/CourseContentType.java
@@ -1,4 +1,4 @@
-package com.overhere.backend.entity.enums;
+package com.overhere.backend.domain.enums;
 
 public enum CourseContentType {
     //코스 카테고리 정해지면 추가
