--- src/main/java/com/overhere/backend/entity/enums/AreaCode.java
@@ -1,5 +1,0 @@
-package com.overhere.backend.entity.enums;
-
-public enum AreaCode {
-    TEMP;
-}
